Commit from VS build. Hopefully this works

+++ gaussian_smooth.hpp
@@ -15,7 +15,7 @@ using namespace arma;
 mat g_smooth (mat img, double scale){
 	//int sigma = 1;
 	int mask_size = 100;
-	float thresh = 1e-3; //THIS IS CORRECT
+	double thresh = 1e-3; //THIS IS CORRECT
 	
 	double scale_factor = (1 / scale);
